fix: address feedback regarding refactors
