+++ pkg/gateway/routes.go
@@ -226,41 +226,3 @@ func isHostnameSubset(routeHostname, listenerHostname string) bool {
 
 	return false
 }
-
-// hostnamesOverlap reports whether two hostname lists share any intersecting name.
-func hostnamesOverlap(a, b []string) bool {
-	for _, ha := range a {
-		for _, hb := range b {
-			if hostnamesIntersect(ha, hb) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func hostnamesIntersect(a, b string) bool {
-	if a == "*" || b == "*" || a == b {
-		return true
-	}
-	return isHostnameSubset(a, b) || isHostnameSubset(b, a)
-}
-
-func routeHostnamesOverlap(listener gatewayv1.Listener, aHostnames, bHostnames []gatewayv1.Hostname) bool {
-	return hostnamesOverlap(
-		getIntersectingHostnames(listener, aHostnames),
-		getIntersectingHostnames(listener, bHostnames),
-	)
-}
-
-// routePrecedenceWins reports whether a should be preferred over b when two
-// routes of different types compete for the same hostname. Oldest creation
-// timestamp wins, then "{namespace}/{name}" alphabetical order.
-func routePrecedenceWins(a, b metav1.ObjectMeta) bool {
-	if !a.CreationTimestamp.Equal(&b.CreationTimestamp) {
-		return a.CreationTimestamp.Before(&b.CreationTimestamp)
-	}
-	aKey := a.Namespace + "/" + a.Name
-	bKey := b.Namespace + "/" + b.Name
-	return aKey < bKey
-}
